refactor: remove premature optimization
